backend/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
mysqlclient==1.4.5
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
django-extensions==2.2.5
celery==4.4.2
django-celery-results==1.2.1
black==19.10b0
flake8==3.8.2
numpy==1.19.0
django-model-utils==4.0.0
